+++ src/main/java/com/example/androidserver/web/Controller/HelloController.java
@@ -1,4 +1,4 @@
-package com.example.androidserver.Controller;
+package com.example.androidserver.web.Controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
